feat/RSS-ECOMM-2_27: refactor header actions
