--- src/pages/Layout/components/HeaderActions/header-action.model.ts
@@ -1,13 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface HeaderAction {
-  key: string;
-  to: string;
-  startIcon: ReactNode;
-  text: string;
-  onClick: () => void;
-  buttonStyle: object;
-  menuItemStyle: object;
-  show: boolean;
-  menuOrder: number;
-}
